I did the memory thingy with the map Also I renamed opcode to doorNumber in Register and removed the folder as to makes packaging easier Beyond that, the only thing left is the mark I left on Main
