+++ src/Main.java
@@ -7,6 +7,7 @@ class Main
 	public static void main(String[] args)
 	{
 		System.out.print("Isa esteve aqui!!!!");
+		System.out.println("\nMe too");
 		MainFrame mainFrame = new MainFrame();
 	}
 }
