+++ src/Memory.java
@@ -1,0 +1,16 @@
+package src;
+import java.util.TreeMap;
+
+public class Memory{
+    public TreeMap<Integer,Integer> map;
+    public Memory(){
+        this.map = new TreeMap<Integer,Integer>();
+    }
+
+    /*
+    Para usar o mapa:
+    map.put(int endereco, int conteudo) - coloca o conteudo no endereco(chave do mapa) e retorna o que estava antes (null se nao tinha nada)
+    map.get(int endereco) - retorna o que tem no endereco
+    map.containsKey(int endereco) - true ou false se tiver ou nao algo no endereco
+    */
+}
